backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
python-dotenv==1.0.0
anthropic==0.18.1
google-generativeai==0.3.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
redis==5.0.1
# 数据库相关
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
